Handle onDrop within TerminalInstance (#123329)
